Add biration test snippet
This is a portion of code I wrote with product-specific references removed. The purpose of this test code is to collect ambient acceleration data to create a baseline average, then wait for a prolonged spike in acceleration to start collecting data, and then continue collecting elevated data until a corresponding fall in the average acceleration is detected. This was done to replace a test which only collected the maximum acceleration data.